Organizing the project, defining the reducer and the StoreProvider component
The project was separated in three folders, one for the components, another to the context management, and another to implement the service logic. The reducer basic structure was definined as a set of 7 cases: to get notes/categories (2), to add a note/category (2), to delete a note/category (2), and to update a note (1). Lastly, the StoreProvider component was created, setting an initial state with two empty list, one for the category objects and the other for the note objects