Implementation of list of categories and notes
Categories are listed using dropdowns. Each category dropdown displays a table with the related notes. The categories are being fetched from the DB and then are being stored in the context state. When a category dropdown is clicked, the notes related with the category are fetched before the table is rendered