Add Routing, About and NotFound pages, final refactoring and styling
Routing implementation using react-router-dom and addition of some pages: one for the to do list management, other for the about page and another for a not found path. A final refactoring and styling were implemented, and also some blocks of code were commented to make the code easier to read (I hope so :)).